+++ src/Plainion.GraphViz/Properties/AssemblyInfo.cs
@@ -5,15 +5,15 @@ using System.Reflection;
 [assembly: AssemblyProductAttribute("Plainion.GraphViz")]
 [assembly: AssemblyDescriptionAttribute("Plainion.GraphViz")]
 [assembly: AssemblyCopyrightAttribute("Copyright @ 2019")]
-[assembly: AssemblyVersionAttribute("3.1")]
-[assembly: AssemblyFileVersionAttribute("3.1")]
+[assembly: AssemblyVersionAttribute("3.2")]
+[assembly: AssemblyFileVersionAttribute("3.2")]
 namespace System {
     internal static class AssemblyVersionInformation {
         internal const System.String AssemblyTitle = "Plainion.GraphViz";
         internal const System.String AssemblyProduct = "Plainion.GraphViz";
         internal const System.String AssemblyDescription = "Plainion.GraphViz";
         internal const System.String AssemblyCopyright = "Copyright @ 2019";
-        internal const System.String AssemblyVersion = "3.1";
-        internal const System.String AssemblyFileVersion = "3.1";
+        internal const System.String AssemblyVersion = "3.2";
+        internal const System.String AssemblyFileVersion = "3.2";
     }
 }